requirements.txt
numpy
pandas
opencv-python
matplotlib
pycocotools
torch
torchvision
tqdm
gradio
Pillow
